requirements.txt
# Core dependencies
python-binance>=1.0.19
python-dotenv>=1.0.0
pydantic>=2.0.0

# Data processing and ML
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
optuna>=3.4.0
xgboost>=2.0.0

# Configuration and logging
pyyaml>=6.0
joblib>=1.3.0
loguru>=0.7.0
rich>=13.0.0

# Database
pymongo>=4.6.0

# WebSocket support
unicorn-binance-websocket-api>=2.0.0

# Additional utilities
typing-extensions>=4.8.0
